solutionAgent/requirements.txt
langchain
openai
langchain-community 
tavily-python
python-dotenv
gradio